% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/justmapit.R
\name{jmi}
\alias{jmi}
\alias{jmi}
\alias{jmi.default}
\alias{jmi}
\alias{jmi.BasicRaster}
\alias{jmi}
\alias{jmi.Spatial}
\title{Just map it}
\usage{
jmi(x, ...)

\method{jmi}{default}(x = NULL, ...)

\method{jmi}{BasicRaster}(x = NULL, ...)

\method{jmi}{Spatial}(x = NULL, ...)
}
\arguments{
\item{...}{}
}
\description{
Just map it
}
