% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/justmapit-package.R
\docType{package}
\name{justmapit-package}
\alias{justmapit}
\alias{justmapit-package}
\title{justmapit: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Michael Sumner \email{mdsumner@gmail.com}

}
\keyword{internal}
